refactor: move http server to it's own file
